Create FUNDING.yml
I've wanted to try out this feature of Github for a while now, but never found the motivation or justification to get in people's faces and beg for it, or similarly. Now that I've, again, put a substantial amount of effort into the site, I feel comfortable in adding the Sponsor button.

It will also serve to pay the hosting costs (assuming @TheSecEng accepts that).

It may make sense to have an organization-wide FUNDING.yml, but since the other projects in there are basically non-significant, we can do that later if needed.