Use a single VPTeamMembers component
Here, we use max-height and `overflow: hidden` to display at most 2 rows
of contributors, regardless of the page width. For this, we explicitly
need to set a height for all items.

Since we don't know ahead of time how many those will be, we might as
well "render" everything at once, even though it will be hidden. At
least FireFox would download the avatars of all images on the page
regardless, also when `display: none` was used on a parent DOM
element.

Also make the "button" to expand the list show a pointer.