Resolve input handlers gist
By including the relevant source code inside the repository directly, it
will always be up to date with the guide and does not require updating
separately.

This is made possible using the feature to [import code snippets] and
using [DownGit] to provide a download archive of a repo's
sub-directory. (Other options required inserting a GitHub token, which
was not plug-and-play enough for me.)

[import code snippets]: https://vitepress.dev/guide/markdown#import-code-snippets
[DownGit]: https://github.com/MinhasKamal/DownGit