Rewrite glossary entirely
Also use the custom <Term> tag throughout the entire documentation texts
because the `:term:text_with_underscores:` syntax is inferior to simply
including the text (with a slot).

With these changes, there is also little to no resemblance to the
vuepress glossary plugin that was used before, meaning we should not
violate any license anymore.