Override esbuild version
Satisfies a security vulnerability noted by `pnpm audit`. We may or may
not want to continue overriding recursive dependencies like this, as
long as the overrides don't get out of hand. Let's hope I won't forget
about this, though it appears as if the overrides only apply
conditionally.

Otherwise we'd rather not require `pnpm audit` to pass for our builds.
Or only make it check direct dependencies.