fix:  remove inaccurate statement...
If the override files involve JSON data structures such as
Menus and Keymaps, then the override is a "merge" type
override where the user only needs to modify the PART of the
file being changed (while maintaining the structure).

If the user understands this, then it is already obvious that
files like Plugins cannot be "merged" that way, and so of
course, in those cases, whole files are overridden.